third_party/blink/web_tests/fast/css/min-max-width-expected.txt

Tests offsetWidth using min-width and max-width

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS testWidth("max-width:150px") is 100
PASS testWidth("max-width:50px") is 50
PASS testWidth("min-width:150px") is 150
PASS testWidth("min-width:50px") is 100
PASS successfullyParsed is true

TEST COMPLETE
